requirements.txt
flask
pandas
psycopg2-binary
python-dotenv
requests
sqlalchemy